Fix2.5/drivingdistance

+++ sql/deprecated/deprecated-comments.sql
@@ -41,7 +41,7 @@ COMMENT ON FUNCTION pgr_bdDijkstra( TEXT, INTEGER, INTEGER, BOOLEAN, BOOLEAN)
 COMMENT ON FUNCTION pgr_dijkstra(TEXT, INTEGER, INTEGER, BOOLEAN, BOOLEAN)
     IS 'pgr_dijkstra(Deprecated signature)';
 
-COMMENT ON FUNCTION pgr_drivingDistance(text,  INTEGER,  FLOAT8,  BOOLEAN,  BOOLEAN)
+COMMENT ON FUNCTION pgr_drivingDistance(text,  BIGINT,  FLOAT8,  BOOLEAN,  BOOLEAN)
     IS 'pgr_drivingDistance(Deprecated signature)';
 
 ------------------------
